Revert " refactor(core/biblio): split into pre-processing and rendering parts (#1712)"
This reverts commit 0998305e6a463c617471dc01cdcf82abc9d89c33.
